+++ NEWS.md
@@ -1,3 +1,10 @@
+# wikilake 0.5
+
+* Bug fixes related to non-standard units
+* Bug fixes related to coordinate parsing
+* Added `lake_clean` function and `clean` argument to `lake_wiki` to standardize output units
+* Added ability to pass a vector of names to lake_wiki (#16)
+
 # wikilake 0.4
 
 * Improved handling and standardization of measurement units
@@ -12,6 +19,3 @@
 
 * Bug fixes related to unit detection and parsing
 * Added a `NEWS.md` file to track changes to the package.
-
-
-